feat: add squaredSum and basic tests.
