+++ src/__tests__/meanSquaredError.test.ts
@@ -1,7 +1,7 @@
 import { Matrix } from 'ml-matrix';
 import { describe, expect, it } from 'vitest';
 
-import { meanSquaredError } from '../squaredSum';
+import { meanSquaredError } from '../meanSquaredError';
 
 describe('meanSquaredError', () => {
   const A = new Matrix([
